Add ability conductor size less than 1, thank Timbul Widodo
git-svn-id: svn+ssh://svn.tuxfamily.org/svnroot/qet/qet/trunk@4605 bfdf4180-ca20-0410-9c96-a3a8aa849046

+++ sources/conductorproperties.cpp
@@ -296,7 +296,7 @@ void ConductorProperties::fromXml(QDomElement &e)
 	m_function           = e.attribute("function");
 	m_tension_protocol   = e.attribute("tension-protocol");
 	text_size            = e.attribute("numsize", QString::number(9)).toInt();
-	cond_size            = e.attribute("condsize", QString::number(1)).toInt();
+	cond_size            = e.attribute("condsize", QString::number(1)).toDouble();
 	m_show_text          = e.attribute("displaytext", QString::number(1)).toInt();
 	m_one_text_per_folio = e.attribute("onetextperfolio", QString::number(0)).toInt();
 	verti_rotate_text    = e.attribute("vertirotatetext").toDouble();
